dndR: Dungeons & Dragons Functions for Players and Dungeon Masters

The goal of 'dndR' is to provide a suite of Dungeons & Dragons related functions. This package is meant to be useful both to players and Dungeon Masters (DMs). All functions currently focus on Fifth Edition (a.k.a. "5e") but once the next edition is published functions will likely be expanded to include any rule changes.

See also

Author

Maintainer: Nicholas Lyon njlyon@alumni.iastate.edu (ORCID) (https://njlyon0.github.io/) [copyright holder]

Authors:

Tim Schatto-Eckrodt (https://kudusch.de/)

Other contributors:

Humberto Nappo (ORCID) [contributor]