Roll for All Ability Scores

Rolls for six ability scores using the desired method of rolling (4d6 drop lowest, 3d6, or 1d20). Doesn't assign abilities to facilitate player selection of which score should be each ability for a given character. Prints a warning if the total of all abilities is less than 70 or if any one ability is less than 8.

ability_scores(method = "4d6", quiet = FALSE)

Arguments

method

(character) string of "4d6", "3d6", or "1d20" ("d20" also accepted). Enter your preferred method of rolling for each ability score ("4d6" drops lowest before summing)

quiet

(logical) whether to print warnings if the total score is very low or one ability score is very low

Value

a dataframe of two columns and six rows

Examples

# Roll ability scores using four d6 and dropping the lowest
ability_scores(method = "4d6")
#>   ability score
#> 1      V1    12
#> 2      V2    15
#> 3      V3    11
#> 4      V4    14
#> 5      V5    10
#> 6      V6    10

# Roll using 3d6 and dropping nothing
ability_scores("3d6")
#> Total score low. Consider re-rolling?
#> At least one ability very low. Consider re-rolling?
#>   ability score
#> 1      V1     7
#> 2      V2     9
#> 3      V3    12
#> 4      V4    12
#> 5      V5     8
#> 6      V6    11

# Or if you're truly wild, just roll a d20 for each ability
ability_scores('d20')
#> At least one ability very low. Consider re-rolling?
#>   ability score
#> 1      V1    18
#> 2      V2     7
#> 3      V3    17
#> 4      V4    20
#> 5      V5     6
#> 6      V6    14