Retrieve Full Creature Description Text by Creature Name

Accepts user-provided Dungeons & Dragons creature name(s) and returns the full set of fifth edition creature information and the complete description text (for the 2014 version). Unlike `dndR::creature_list`, this function requires an exact match between the user-provided creature name(s) and how they appear in the main creature data object. The argument in this function is not case-sensitive.

Usage

creature_text(name = NULL, ver = "2014")

Arguments

name

(character) exact creature name(s) for which to gather description information

ver

(character) which version of fifth edition to use ("2014" or "2024"). Note that only 2014 is supported and entering "2024" will print a warning to that effect

Value

(dataframe) one column per creature specified by the user. Creature name is stored as the column name for that creature's information. Returns all fields for which there are data for at least one of the specified creatures so row number will vary with query (maximum 26 rows if all fields have information).

Examples

dndR::creature_text(name = c("hill giant", "goblin"))
#>                                                                                                                                                                                                                         Goblin
#> 1                                                                                                                                                                                                       Source: Monster Manual
#> 2                                                                                                                                                                                                                  STR: 8 (-1)
#> 3                                                                                                                                                                                                                 DEX: 14 (+2)
#> 4                                                                                                                                                                                                                  CON: 10 (0)
#> 5                                                                                                                                                                                                                  INT: 10 (0)
#> 6                                                                                                                                                                                                                  WIS: 8 (-1)
#> 7                                                                                                                                                                                                                  CHA: 8 (-1)
#> 8                                                                                                                                                                                                                  Size: small
#> 9                                                                                                                                                                                                   Type: humanoid (goblinoid)
#> 10                                                                                                                                                                                                     Alignment: neutral evil
#> 11                                                                                                                                                                                                                      XP: 50
#> 12                                                                                                                                                                                                                    CR: 0.25
#> 13                                                                                                                                                                                                   Languages: common, goblin
#> 14                                                                                                                                                                                                          Skills: Stealth +6
#> 15                                                                                                                                                                                                               Speed: 30 ft.
#> 16                                                                                                                                                                                                                 HP: 7 (2d6)
#> 17                                                                                                                                                                                              AC: 15 (leather armor, shield)
#> 18                                                                                                                                                                                                           darkvision 60 ft.
#> 19                                                                                                               ***Nimble Escape.*** The goblin can take the Disengage or Hide action as a bonus action on each of its turns.
#> 20 ***Scimitar.*** Melee Weapon Attack: +4 to hit, reach 5 ft., one target. Hit: 5 (1d6 + 2) slashing damage. ***Shortbow.*** Ranged Weapon Attack: +4 to hit, range 80/320 ft., one target. Hit: 5 (1d6 + 2) piercing damage.
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            Hill Giant
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              Source: Monster Manual
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        STR: 21 (+5)
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         DEX: 8 (-1)
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        CON: 19 (+4)
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         INT: 5 (-3)
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         WIS: 9 (-1)
#> 7                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         CHA: 6 (-2)
#> 8                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          Size: huge
#> 9                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         Type: giant
#> 10                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            Alignment: chaotic evil
#> 11                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           XP: 1800
#> 12                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              CR: 5
#> 13                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   Languages: giant
#> 14                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              Skills: Perception +2
#> 15                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      Speed: 40 ft.
#> 16                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 HP: 105 (10d12+40)
#> 17                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             AC: 13 (natural armor)
#> 18                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               <NA>
#> 19                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               <NA>
#> 20 ***Greatclub.*** Melee Weapon Attack: +8 to hit, reach 10 ft., one target. Hit: 18 (3d8 + 5) bludgeoning damage. ***Multiattack.*** The giant makes two greatclub attacks. ***Rock.*** Ranged Weapon Attack: +8 to hit, range 60/240 ft., one target. Hit: 21 (3d10 + 5) bludgeoning damage. ***Squash.*** Melee Weapon Attack: +8 to hit, reach 5 ft., one Medium or Smaller creature. Hit: 26 (6d6 + 5) bludgeoning damage, the giant lands prone in the target's space, and the target is grappled (escape DC 15). Until this grapple ends, the target is prone. The grapple ends early if the giant stands up.