Creates a Monster for Given Party Level and Size

Returns the vital statistics of a randomized monster based on a the average player level and number of players in the party. This function follows the advice of [Zee Bashew](https://twitter.com/Zeebashew) on how to build interesting, challenging monsters for your party. These monsters are built somewhat according to the Dungeon Master's Guide for creating monsters, partly Zee's [YouTube video on homebrewing monsters based on The Witcher videogame](https://www.youtube.com/watch?v=GhjkPv4qo5w), and partly on my own sensibilities about scaling the difficulty of a creature. Creatures are spawned randomly so you may need to re-run the function several times (or mentally modify one or more parts of the output) to get a monster that fits your campaign and players, but the vulnerabilities and resistances should allow for cool quest building around what this function provides. Happy DMing!

monster_creator(party_level = NULL, party_size = NULL)

Arguments

party_level

(numeric) indicating the average party level. If all players are the same level, that level is the average party level

party_size

(numeric) indicating how many player characters (PCs) are in the party

Value

(dataframe) two columns and 15 rows

Examples

# Creates a monster from the specified average party level
monster_creator(party_level = 4, party_size = 3)
#>             statistic                                     value
#> 1          Hit_Points                                        77
#> 2         Armor_Class                                        15
#> 3          Prof_Bonus                                         3
#> 4        Attack_Bonus                                         6
#> 5             Save_DC                                        15
#> 6  Prof_Saving_Throws                                  STR; CON
#> 7           Immune_to                          psychic; radiant
#> 8        Resistant_to acid; lightning; necrotic; fire; piercing
#> 9       Vulnerable_to                                      cold
#> 10                STR                                        +3
#> 11                DEX                                        +2
#> 12                CON                                        +2
#> 13                INT                                        +2
#> 14                WIS                                        +2
#> 15                CHA                                        +1